✨ Extend help system
Added new HelpCommand class and changed CommandContainer from an
interface to a class to support new help system.

This new help removes a nasty circular dependency in src/commands and
prevents needing to reimplement help for every new collection of
subcommands (like the set of mod commands currently in progress).