🚚 Delete utils directory and move its contents
Instead of splitting utils into separate files in its own directory,
I've moved it into utils.ts and moved some of the other files/functions
out of utils entirely.

I'm still not entirely sure what to do with utils. It's basically a
place I put things that I don't yet know a better place for. 🤔

But this is probably a step in a better direction for now.